backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
alembic==1.12.1
pyjwt==2.8.0
google-auth==2.23.4
python-jose[cryptography]==3.3.0